requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
pygame>=2.5.0
